requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
faiss-cpu>=1.7.4
openai>=1.3.7
python-dotenv>=1.0.0
tiktoken>=0.5.2
numpy>=1.24.0
pandas>=2.0.0 